webapp/requirements.txt
fastapi
uvicorn[standard]
PyYAML
python-multipart
websockets
python-hcl2